+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add abstract `WithDataprovider` class and make `al-table` and `al-select` use it
 * Allow `al-table` to reuse existing rows and prevent render flashes
 * Add `al-state` and the abstract `Stateful` class for elements with states
+* Add `al-button` element
 
 ## 0.1.12 (2022-10-12)